# Remove String conversion in decoders.