Proposed fix for #511. The only thing I was unsure about was whether to use `Buf.ByteArray.Owned` or `Buf.ByteArray.Shared` for the libraries that can convert `Array[Byte]`. I wound up using `Shared`, but I'm not sure this is necessary.