@vkostyukov 2016-11-19
@rpless Thanks a lot for looking into this!

Some suggestions:

 - Instead of `Buf.ByteArray.Shared` (which always copy), let's use [this trick](https://github.com/finagle/finch/blob/master/core/src/main/scala/io/finch/Endpoints.scala#L287-L292) that allows retrieving a byte array from Netty's channel buffer. This heavily depends on the underlying implementation, but I'm fine with that for now.
 - Let run some benchmarks (both JMH and wrk) to see how this affects throughput (I'm excited to see the numbers!)