@vkostyukov 2016-11-19
Trying to be more specific here.

If you need `ByteBuffer`, the cheapest way (at least for Netty 3) to get it is:

```scala
val byteBuffer = ChannelBufferBuf.Owned.extract(buf).toByteBuffer()
```

If you need `Array[Byte]`, the cheapest way to get it is:

```scala
val channelBuffer = ChannelBufferBuf.Owned.extract(buf)
val (byteArray, offset, length) = 
  // assert channelBuffer.hasArray()
  (channelBuffer.array(), channelBuffer.readerIndex(), channelBuffer.readableBytes()) 
```