@rpless 2016-11-19
I ran the wrk benchmarks for Finch + circe. The results are [here](https://gist.github.com/rpless/7149d0f110e25a9885be4fa8582c6a55).
As for jmh benchmarks, unless I missed something, it doesn't look like we have ones that exercise the the json decoders anymore.