@vkostyukov 2016-11-19
Nice! 

@rpless Do you mind keep the server running and run `wrk` for 3 times. Usually, it takes some time for JVM to warmup so using the results from 3rd run seems like the closest we can get to the steady state.