@rpless 2016-11-19
Updated benckmarks, results after 3 runs:
## Without String Conversion (this PR)
```
$ wrk -t4 -c24 -d30s -s examples/src/main/scala/io/finch/wrk/wrk.lua http://localhost:8081/
Running 30s test @ http://localhost:8081/
  4 threads and 24 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.68ms    3.67ms  52.43ms   90.18%
    Req/Sec    12.20k     2.26k   23.58k    67.25%
  1458406 requests in 30.09s, 137.69MB read
Requests/sec:  48471.67
Transfer/sec:      4.58MB
```

## With String Conversion (Current Finch)
```
$ wrk -t4 -c24 -d30s -s examples/src/main/scala/io/finch/wrk/wrk.lua http://localhost:8081/
Running 30s test @ http://localhost:8081/
  4 threads and 24 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.62ms    3.22ms  62.64ms   89.91%
    Req/Sec    10.84k     2.44k   22.20k    68.33%
  1296147 requests in 30.10s, 122.37MB read
Requests/sec:  43067.06
Transfer/sec:      4.07MB
```