@vkostyukov 2016-11-19
Looks like **12.5%** improvement in the throughput! Sweet!