@vkostyukov 2016-11-19
I think Chris is right. We call Json.parse two times and drop the first
result on the floor.

On Sat, Nov 19, 2016 at 11:14 AM Ryan Plessner <notifications@github.com>
wrote:

> *@rpless* commented on this pull request.
> ------------------------------
>
> In playjson/src/main/scala/io/finch/playjson/package.scala
> <https://github.com/finagle/finch/pull/671>:
>
> > @@ -11,10 +13,11 @@ package object playjson {
>     * @tparam A the type of the data to decode into
>     */
>    implicit def decodePlayJson[A](implicit reads: Reads[A]): Decode.Json[A] =
> -    // TODO: Eliminate toString conversion
> -    // See https://github.com/finagle/finch/issues/511
> -    // PlayJson can parse from byte[] automatically detecting the charset.
> -    Decode.json((b, cs) => Try(Json.parse(BufText.extract(b, cs)).as[A]))
> +    Decode.json({ (b, cs) => Try(Json.parse(Buf.ByteArray.Shared.extract(b)).as[A])
> +      val buf = ChannelBufferBuf.Owned.extract(b)
> +      if (buf.hasArray) Try(Json.parse(new String(buf.array(), 0, buf.readableBytes(), cs)).as[A])
> +      else Try(Json.parse(buf.toString(cs)).as[A])
> +    })
>
> @clhodapp <https://github.com/clhodapp> I played around with it a little.
> Do you think this version makes the intent more clear?
>
> val buf = ChannelBufferBuf.Owned.extract(b)val bufAsString = Try(
>   if (buf.hasArray) new String(buf.array(), 0, buf.readableBytes(), cs)
>   else buf.toString(cs)
> )
> bufAsString.map(Json.parse(_).as[A])
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/finagle/finch/pull/671>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABDDKwEmt9Qm3tSF-W2xoHHcsjNTIvoyks5q_0p5gaJpZM4K3Isa>
> .
>