@rpless 2016-11-19
Ok, it looks like using `slice` like this works: `buf.array().slice(0, buf.readableBytes())`. Does that make sense?