@vkostyukov 2016-11-19
Slice works but it allocates a new array. You should be able to bypass
readableBytes to the parse function.

On Sat, Nov 19, 2016 at 12:48 PM Ryan Plessner <notifications@github.com>
wrote:

> Ok, it looks like using slice like this works: buf.array().slice(0,
> buf.readableBytes()). Does that make sense?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/finagle/finch/pull/671#issuecomment-261738691>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ABDDK-0w2n0qAoIwWMcWBZ8h16aUD1ooks5q_2CrgaJpZM4K3Isa>
> .
>