@rpless 2016-11-19
Yeah, I'm not a fan of the new array allocation, but it seems like neither Spray Json nor Play Json expose a parsing function that takes the number of bytes to read:
- https://www.playframework.com/documentation/2.5.x/api/scala/index.html#play.api.libs.json.Json$
- https://github.com/spray/spray-json/blob/master/src/main/scala/spray/json/JsonParser.scala#L28
  (and also https://github.com/spray/spray-json/blob/master/src/main/scala/spray/json/JsonParser.scala#L237)