@vkostyukov 2016-11-19
Play takes InputStream though. Can we wrap our byte array with
ByteArrayInputStream and specify the length there?

For Spray, let's do System.arraycopy it's still better than 'new String'.

On Sat, Nov 19, 2016 at 1:10 PM Ryan Plessner <notifications@github.com>
wrote:

> Yeah, I'm not a fan of the new array allocation, but it seems like neither
> Spray Json nor Play Json expose a parsing function that takes the number of
> bytes to read:
>
>    -
>    https://www.playframework.com/documentation/2.5.x/api/scala/index.html#play.api.libs.json.Json$
>    -
>    https://github.com/spray/spray-json/blob/master/src/main/scala/spray/json/JsonParser.scala#L28
>    (and also
>    https://github.com/spray/spray-json/blob/master/src/main/scala/spray/json/JsonParser.scala#L237
>    )
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/finagle/finch/pull/671#issuecomment-261739768>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ABDDK_tX58WsqNgUMHNRIP01LhdK2MDxks5q_2WagaJpZM4K3Isa>
> .
>