@rpless 2016-11-19
Can do for Spray, but it looks like I gave you the wrong docs for Play Json. We're currently on [2.3.x](https://www.playframework.com/documentation/2.3.x/api/scala/index.html#play.api.libs.json.Json$) which does not provide a parse with InputStream. So we can either bump the Play version or System.arraycopy.