@vkostyukov 2016-11-19
Arraycopy works for me. Let's update it later.

On Sat, Nov 19, 2016 at 1:50 PM Ryan Plessner <notifications@github.com>
wrote:

> Can do for Spray, but it looks like I gave you the wrong docs for Play
> Json. We're currently on 2.3.x
> <https://www.playframework.com/documentation/2.3.x/api/scala/index.html#play.api.libs.json.Json%24>
> which does not provide a parse with InputStream. So we can either bump the
> Play version or System.arraycopy.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/finagle/finch/pull/671#issuecomment-261742009>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ABDDK22ipWPepiBj6q_IqnkB5M6wpEX7ks5q_28_gaJpZM4K3Isa>
> .
>