@vkostyukov 2016-11-21
Thanks again @rpless! Merging this.